Moved values state from RecordsOptions to RecordsView.
